backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.1
pydantic==2.5.0
python-chess==1.999
python-dotenv==1.0.0
mcp==1.0.0
openai>=1.0.0

